+++ src/api.ts
@@ -187,7 +187,6 @@ export interface ContractCallQuoteRequest extends ToolConfiguration {
   referrer?: string
   fee?: number | string
   allowDestinationCall?: boolean // (default : true) // destination calls are enabled by default
-  maxPriceImpact?: number // hide routes with price impact greater than or equal to this value
 }
 
 export interface ContractCallQuotesRequest extends ToolConfiguration {
@@ -426,16 +425,20 @@ export type RefetchLIFuelRequest = {
   chainId: ChainId
 }
 
-export type GasRecommendationResponse = {
-  available: boolean // whether we can support that
-  recommended?: TokenBalance
-  limit?: TokenBalance // Maximum of gas the user can transfer
-  serviceFee?: TokenBalance // LI.FI fee for providing the service
+export type GasRecommendationResponse =
+  | {
+      // whether we can support that
+      available: false
+      // reason why the gas feature is not available (e.g. missing liquidity)
+      message: string
+    }
+  | {
+      available: true
+      recommended: TokenBalance
+      limit: TokenBalance // Maximum of gas the user can transfer
+      serviceFee: TokenBalance // LI.FI fee for providing the service
 
-  // information about what the user has to pay to get the recommended gas amount
-  fromToken?: Token
-  fromAmount?: string
-
-  // if available is false
-  message?: string // reason why the gas feature is not available (e.g. missing liquidity)
-}
+      // information about what the user has to pay to get the recommended gas amount
+      fromToken?: Token
+      fromAmount?: string
+    }